Update Dockerfile
The primary change in this case is that we are preparing the build to no longer case about using docker within docker.  This is because (1) docker actually runs slower inside docker due to the use of creating new volumes and (2) those volumes do not get removed upon exiting the container.  

- restructure build commands so that it'll take better advantage of caching.
- install newest version of docker
- only install wget.  Removed installation of curl as it takes up more space
- set default entry point to the docker wrapper
- set default command to launch jenkins